Add tests for Windows CLI discovery, CRLF parsing, and deep-link cleanup
Moves the pure parsers into parse.ts and the per-platform search config into
platform.ts so they can be exercised with Node's built-in test runner without
the Raycast runtime. Adds `npm test` with CRLF fixtures and %0D%0A link cases.
