+++ extensions/spark-mail/package.json
@@ -82,6 +82,7 @@
     "dev": "ray develop",
     "lint": "ray lint",
     "fix-lint": "ray lint --fix",
+    "test": "node --test src/lib/spark.test.ts",
     "publish": "npx @raycast/api@latest publish"
   }
 }
